Add rest controller api documentation #35
Just started with it, not done yet.

+++ server/src/main/java/com/plantstein/server/model/Species.java
@@ -18,7 +18,7 @@ public class Species {
 private String name;
 
 @NotEmpty
-private int perfectTemperature;
+private double perfectTemperature;
 
 @NotEmpty
 private double perfectLight;
